Fix: mealsコントローラーのリファクタリング
