+++ app/controllers/meals_controller.rb
@@ -1,13 +1,10 @@
 class MealsController < ApplicationController
-  require "google/cloud/vision"
-  require "google/cloud/translate/v2"
-
   def index
     @meals = Meal.with_result.includes(:foods, :user).order(created_at: :desc).page(params[:page]).per(24)
   end
 
   def show
-    @meal = Meal.all.includes(:foods, :user).find(params[:id])
+    @meal = Meal.find(params[:id])
     @data_values = [@meal.user.sugar_limit.round, @meal.sugar_intake]
   end
   
@@ -20,7 +17,7 @@ class MealsController < ApplicationController
         @meal = current_user.meals.build(meal_params)
       if @meal.save
         sent_image = File.open(meal_params["meal_image"].tempfile)
-        @meal.update!(analyzed_foods: image_analysis(sent_image))
+        @meal.update!(analyzed_foods: @meal.image_analysis(sent_image))
         redirect_to edit_meal_path(@meal), success: t('.success')
       else
         flash.now[:danger] = t('.failure')
@@ -66,38 +63,5 @@ class MealsController < ApplicationController
   
   def meal_params
     params.require(:meal).permit(:meal_image)
-  end
-
-  def image_analysis(meal_image)
-    File.open("app/google-credentials.json", 'w') do |file|
-      if Rails.env.production?
-        JSON.dump(Rails.application.credentials.production_google, file)
-      else
-        JSON.dump(Rails.application.credentials.development_google, file)
-      end
-    end
-
-    ENV["GOOGLE_APPLICATION_CREDENTIALS"] = "app/google-credentials.json"
-  
-    image_annotator = Google::Cloud::Vision.image_annotator
-    translate = Google::Cloud::Translate::V2.new
-
-
-    response = image_annotator.label_detection(
-      image:   meal_image,
-    )
-
-    results = []
-
-    response.responses.each do |res|
-      res.label_annotations.each do |label|
-        translation = translate.translate label.description.downcase, to: 'ja'
-        results << translation.text
-      end
-    end
-    
-    File.delete("app/google-credentials.json")
-
-    results.join(',')
   end
 end
